inherit base CLI class for tasks
